// frontend/src/api/model/resumeResponse.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Libre-Q API
 * OpenAPI spec version: 0.1.0
 */
import type { ResumeResponseDraftResponses } from './resumeResponseDraftResponses';

/**
 * Schema for returning participant resume data.
 */
export interface ResumeResponse {
    session_token: string;
    language: string;
    last_step_reached: number;
    draft_responses: ResumeResponseDraftResponses;
}